several warnings corrected - related to 32/64 bit windows - big chagnes to IME API (fixes) for Led, and a few other casts, changes of type for Win64

+++ Library/Sources/Stroika/Frameworks/Led/OptionsSupport.cpp
@@ -242,7 +242,7 @@ void OptionsFileHelper::StorePref (const Led_SDK_Char* prefName, const string& v
 #if qPlatform_MacOS
     Assert (false); // NYI
 #elif qPlatform_Windows
-    (void)::RegSetValueExA (fKey, Led_SDKString2ANSI (prefName).c_str (), 0, REG_SZ, (LPBYTE)value.c_str (), (value.length () + 1) * sizeof (TCHAR));
+    (void)::RegSetValueExA (fKey, Led_SDKString2ANSI (prefName).c_str (), 0, REG_SZ, (LPBYTE)value.c_str (), static_cast<DWORD> ((value.length () + 1) * sizeof (TCHAR)));
 #endif
 }
